Lib

Documentation

foldLoop' :: Int -> (a -> Int -> IO a) -> a -> IO a

Swaps the arguments of HaskTorch's foldLoop around

nanToNum :: Float -> Float -> Float -> Tensor -> Tensor

Because snake_case sucks

nanToNum' :: Tensor -> Tensor

Default limits for nanToNum

nanToNum'' :: Tensor -> Tensor

Default limits for nanToNum (0.0)

fullLike' :: Tensor -> Float -> Tensor

GPU Tensor filled with Float value

indexSelect'' :: Int -> [Int] -> Tensor -> Tensor

Select index with [Int] from GPU tensor

weightLimit :: Linear -> Float

Calculate weight Limits based on Layer Dimensions

weightInit :: Float -> Linear -> IO Linear

Initialize Weights of Linear Layer

weightInit' :: Linear -> IO Linear

Initialize weights based on Fan In

softUpdate :: Tensor -> Tensor -> Tensor -> Tensor

Softly update parameters from Online Net to Target Net

softSync :: Parameterized f => Tensor -> f -> f -> IO f

Softly copy parameters from Online Net to Target Net

copySync :: Parameterized f => f -> f -> f

Hard Copy of Parameter from one net to the other

gpu :: Device

GPU 1

cpu :: Device

CPU 0

dataType :: DType

Default Tensor Data Type

toTensor :: TensorLike a => a -> Tensor

Convert an Array to a Tensor

toIntTensor :: TensorLike a => a -> Tensor

Convert an Array to a Tensor

emptyTensor :: Tensor

Create an empty Float Tensor on GPU

toScalar :: Float -> Tensor

Convert a Scalar to a Tensor

toDoubleGPU :: forall a. HasTypes a Tensor => a -> a

Convert model to Double on GPU

toFloatGPU :: forall a. HasTypes a Tensor => a -> a

Convert model to Float on CPU

toFloatCPU :: forall a. HasTypes a Tensor => a -> a

Convert model to Float on CPU

randomInts :: Int -> Int -> Int -> IO Tensor

Generate a Tensor of random Integers

normal' :: [Int] -> IO Tensor

Generate Normally Distributed Random values given dimensions

normal :: Tensor -> Tensor -> IO Tensor

Generate Normally Distributed Random values given μs and σs

data Normal

Normal Distribution until HaskTorch brings it's own

Constructors

Normal

Fields

loc :: Tensor

scale :: Tensor

sample :: Normal -> IO Tensor

Sample from Normal Distribution

prob :: Normal -> Tensor -> Tensor

PDF for given Normal Distribution

logProb :: Normal -> Tensor -> Tensor

logarithmic probabilites for given Normal Distribution

entropy :: Normal -> Tensor

Maximum Entropy of Normal Distribution

data Info

Info object gotten form stepping

Constructors

Info

Fields

observations :: ![String]

actions :: ![String]

data Step

Environment Step

Constructors

Step

Fields

observation :: ![Float]

reward :: !Float

done :: !Bool

info :: !Info

type HymURL = String

Base Route to Hym Server

mapToTensor :: Map Int [Float] -> Tensor

Convert a Map to a Tensor where Pool index is a dimension

tensorToMap :: Tensor -> Map Int [Float]

Convert Tensor to Map

stepsToTuple :: Map Int Step -> (Tensor, Tensor, Tensor, [Info])

Convert the Pooled Step Map to a Tuple

hymGet :: HymURL -> String -> IO ByteString

Generic HTTP GET Request to Hym Server

hymPost :: HymURL -> String -> Value -> IO ByteString

Send a POST Request to a Hym Server

hymPoolMap :: HymURL -> String -> IO (Map Int (Map String Float))

Convert a JSON Response from an ACE Server to a Map

hymPoolList :: HymURL -> String -> IO (Map Int [Float])

Convert a JSON Response from an ACE Server to a List

hymPoolReset :: HymURL -> IO (Map Int [Float])

Reset Pooled Environments on a Hym server

hymPoolRandomAction :: HymURL -> IO (Map Int [Float])

Get Random Actions from all Pooled Environments

hymPoolRandomStep :: HymURL -> IO (Map Int Step)

Perform Random Actions in all Pooled Environments

hymPoolStep :: HymURL -> Map Int [Float] -> IO (Map Int Step)

Take Steps in All Environments

aceURL :: String -> String -> String -> String -> String -> HymURL

Generate URL to a Hym-GACE server from meta information

gymURL :: String -> String -> String -> String -> HymURL

Generate URL to a Hym-Gym server from meta information

acePoolTarget :: HymURL -> IO (Map Int (Map String Float))

Send a GET Request to a GACE Server Obtain the Target of Pooled GACE Environments

acePoolActKeys :: HymURL -> IO (Map Int [String])

Action Keys from GACE Server

acePoolObsKeys :: HymURL -> IO (Map Int [String])

Observation Keys from GACE Server

shaceLogPath' :: HymURL -> IO (Map String String)

Get the SHACE logging path as a dict

shaceLogPath :: HymURL -> IO String

Get the SHACE logging path

resetPool :: HymURL -> IO Tensor

Reset a Vectorized Environment Pool

resetPool' :: HymURL -> Tensor -> IO Tensor

Reset selected Environments from Pool

actKeysPool :: HymURL -> IO [String]

Shorthand for getting keys of pooled same envs

obsKeysPool :: HymURL -> IO [String]

Shorthand for getting keys of pooled same envs

infoPool :: HymURL -> IO Info

Get Info without stepping

stepPool :: HymURL -> Tensor -> IO (Tensor, Tensor, Tensor, [Info])

Step in an Environment

randomStepPool :: HymURL -> IO (Tensor, Tensor, Tensor, [Info])

Take a random Step an Environment

randomActionPool :: HymURL -> IO Tensor

Get a set of random actions from the current environment

processGace :: Tensor -> Info -> Tensor

Process / Sanitize the Observations from GACE

scaleRewards :: Tensor -> Tensor -> Tensor

Scale reward to center

createLogDir :: FilePath -> IO ()

Create a log directory

setupLogging :: FilePath -> IO ()

Setup the Logging Direcotry for SHACE

remoteLogPath :: HymURL -> IO FilePath

Get SHACE Logging path to a given URL

writeLoss :: Int -> Int -> String -> Float -> IO ()

Append a line to the give log file

writeReward :: Int -> Int -> Int -> Float -> IO ()

Append a line to the give log file

writeReward' :: Int -> Int -> Tensor -> IO ()

Convenience function for logging an entire env tensor

writeEnvLog :: FilePath -> HymURL -> IO ()

Obtain current performance from a gace server and write/append to log

writeLog :: FilePath -> Map String [Float] -> IO ()

Write an arbitrary Map to a log file for visualization