Lib

Description

Utility and Helper functions for EDELWACE

Documentation

data Algorithm

Deep Reinforcement Learning Algorithm

Constructors

SAC
Soft Actor-Critic

TD3
Twin Delayed Deep Deterministic Policy Gradient

PPO
Proximal Policy Optimization

foldLoop' :: Int -> (a -> Int -> IO a) -> a -> IO a

Swaps the arguments of HaskTorch's foldLoop around

nanToNum :: Float -> Float -> Float -> Tensor -> Tensor

Because snake_case sucks

nanToNum' :: Tensor -> Tensor

Default limits for nanToNum

nanToNum'' :: Tensor -> Tensor

Default limits for nanToNum (0.0)

fullLike' :: Tensor -> Float -> Tensor

GPU Tensor filled with Float value

indexSelect'' :: Int -> [Int] -> Tensor -> Tensor

Select index with [Int] from GPU tensor

repeatInterleave' :: Int -> Tensor -> Tensor -> Tensor

Syntactic sugar for HaskTorch's repeatInterleave so it can more easily be fmapped.

both :: (a -> b) -> Either a a -> b

Apply same function to both Left and Right

both' :: (a -> b) -> Either a a -> Either b b

Apply same function to both Left and Right and put back into Either

range :: Int -> [Int]

Range from 0 to n - 1

fst' :: (a, b, c) -> a

First of triple

splits :: [Int] -> [[Int]]

Helper function creating split indices

splits' :: [Int] -> [Tensor]

Helper function creating split indices as gpu int tensor

currentTimeStamp :: String -> IO String

Current Timestamp as formatted string

currentTimeStamp' :: IO String

Current Timestamp with default formatting: "%Y%m%d-%H%M%S"

createModelArchiveDir :: String -> IO String

Create a model archive directory for the given algorithm

weightLimit :: Linear -> Float

Calculate weight Limits based on Layer Dimensions

data Initializer

Type of weight initialization

Constructors

Normal
Normally distributed weights

Uniform
Uniformally distributed weights

XavierNormal
Using T.xavierNormal

XavierUniform
Using T.xavierUniform

KaimingNormal
Using T.kaimingNormal

KaimingUniform
Using T.kaimingUniform

Dirac

Eye

Ones

Zeros

Constant

initWeights :: Initializer -> Float -> Float -> [Int] -> IO IndependentTensor

Weights for a layer given limits and dimensions.

weightInit :: Initializer -> Float -> Float -> Linear -> IO Linear

Initialize Weights of Linear Layer

weightInitUniform :: Float -> Float -> Linear -> IO Linear

Initialize weights uniformally given upper and lower bounds

weightInitUniform' :: Linear -> IO Linear

Initialize weights uniformally based on Fan In

weightInitNormal :: Float -> Float -> Linear -> IO Linear

Initialize weights normally given mean and std bounds

weightInitNormal' :: Linear -> IO Linear

Initialize weights normally based on Fan In

softUpdate :: Tensor -> Tensor -> Tensor -> Tensor

Softly update parameters from Online Net to Target Net

softSync :: Parameterized f => Tensor -> f -> f -> IO f

Softly copy parameters from Online Net to Target Net

copySync :: Parameterized f => f -> f -> f

Hard Copy of Parameter from one net to the other

gpu :: Device

GPU 1

cpu :: Device

CPU 0

dataType :: DType

Default Tensor Data Type

toTensor :: TensorLike a => a -> Tensor

Convert an Array to a Tensor

toIntTensor :: TensorLike a => a -> Tensor

Convert an Array to a Tensor

emptyTensor :: Tensor

Create an empty Float Tensor on GPU

toScalar :: Float -> Tensor

Convert a Scalar to a Tensor

toDoubleGPU :: forall a. HasTypes a Tensor => a -> a

Convert model to Double on GPU

toFloatGPU :: forall a. HasTypes a Tensor => a -> a

Convert model to Float on CPU

toFloatCPU :: forall a. HasTypes a Tensor => a -> a

Convert model to Float on CPU

randomInts :: Int -> Int -> Int -> IO Tensor

Generate a Tensor of random Integers

normal' :: [Int] -> IO Tensor

Generate Normally Distributed Random values given dimensions

normal :: Tensor -> Tensor -> IO Tensor

Generate Normally Distributed Random values given μs and σs

uniform' :: [Int] -> Float -> Float -> IO Tensor

Generate Uniformally distributed values in a given range

httpOptions :: Options

Default HTTP options for Hym server communication

data Info

Info object gotten form stepping

Constructors

Info

Fields

observations :: ![String]

Observation Keys

actions :: ![String]

Action Keys

data Step

Single Environment Step

Constructors

Step

Fields

observation :: ![Float]

Observation Vector

reward :: !Float

Reward Scalar

done :: !Bool

Terminal Indicator

info :: !Info

Info

type HymURL = String

Base Route to Hym Server

data ActionSpace

Possible Action Spaces

Constructors

Continuous
Continuous Action Space

Discrete
Discrete Action Space

mapToTensor :: Map Int [Float] -> Tensor

Convert a Map to a Tensor where Pool index is a dimension

tensorToMap :: Tensor -> Map Int [Float]

Convert Tensor to Map (Continuous action spaces)

tensorToMap' :: Tensor -> Map Int Int

Convert Tensor to Map (Discrete action spaces)

stepsToTuple :: Map Int Step -> (Tensor, Tensor, Tensor, [Info])

Convert the Pooled Step Map to a Tuple

hymGet :: HymURL -> String -> IO ByteString

Generic HTTP GET Request to Hym Server

hymPost :: HymURL -> String -> Value -> IO ByteString

Send a POST Request to a Hym Server

hymPoolMap :: HymURL -> String -> IO (Map Int (Map String Float))

Convert a JSON Response from an ACE Server to a Map

hymPoolList :: HymURL -> String -> IO (Map Int [Float])

Convert a JSON Response from an ACE Server to a Float-List

hymPoolList' :: HymURL -> String -> IO (Map Int [String])

Convert a JSON Response from an ACE Server to a String-List

hymPoolReset :: HymURL -> IO (Map Int [Float])

Reset Pooled Environments on a Hym server

hymPoolRandomAction :: HymURL -> IO (Map Int [Float])

Get Random Actions from all Pooled Environments

hymPoolRandomStep :: HymURL -> IO (Map Int Step)

Perform Random Actions in all Pooled Environments

hymPoolStep :: HymURL -> Map Int [Float] -> IO (Map Int Step)

Take Steps in All Environments (Continuous)

hymPoolStep' :: HymURL -> Map Int Int -> IO (Map Int Step)

Take Steps in All Environments (Discrete)

aceURL :: String -> String -> String -> String -> String -> HymURL

Generate URL to a Hym-GACE server from meta information

gymURL :: String -> String -> String -> String -> HymURL

Generate URL to a Hym-Gym server from meta information

acePoolTarget :: HymURL -> IO (Map Int (Map String Float))

Send a GET Request to a GACE Server Obtain the Target of Pooled GACE Environments

acePoolActKeys :: HymURL -> IO (Map Int [String])

Action Keys from GACE Server

acePoolObsKeys :: HymURL -> IO (Map Int [String])

Observation Keys from GACE Server

shaceLogPath' :: HymURL -> IO (Map String String)

Get the SHACE logging path as a dict

shaceLogPath :: HymURL -> IO String

Get the SHACE logging path

resetPool :: HymURL -> IO Tensor

Reset a Vectorized Environment Pool

resetPool' :: HymURL -> Tensor -> IO Tensor

Reset selected Environments from Pool

actKeysPool :: HymURL -> IO [String]

Shorthand for getting keys of pooled same envs

obsKeysPool :: HymURL -> IO [String]

Shorthand for getting keys of pooled same envs

numEnvsPool :: HymURL -> IO Int

Number of Environments in Pool

infoPool :: HymURL -> IO Info

Get Info without stepping

targetPool :: HymURL -> IO Tensor

Get Targets for all envs in Pool

targetPool' :: HymURL -> IO Tensor

Get Targets for all envs in Pool and process them

stepPool :: HymURL -> Tensor -> IO (Tensor, Tensor, Tensor, [Info])

Step in a Control Environment

stepPool' :: HymURL -> Tensor -> IO (Tensor, Tensor, Tensor, [Info])

Step in a Discrete Environment

randomStepPool :: HymURL -> IO (Tensor, Tensor, Tensor, [Info])

Take a random Step an Environment

randomActionPool :: HymURL -> IO Tensor

Get a set of random actions from the current environment

saveOptim :: Adam -> FilePath -> IO ()

Optimizer moments at given prefix

loadOptim :: Int -> Float -> Float -> FilePath -> IO Adam

Load Optimizer State

boolMask :: Int -> [Int] -> Tensor

Create Boolean Mask Tensor from list of indices.

boolMask' :: Int -> Tensor -> Tensor

Create a Boolean Mask Tensor from index Tensor

processTarget :: Map Int (Map String Float) -> Tensor

Process Targets for HER

processGace'' :: Tensor -> Info -> (Tensor, Tensor, Tensor)

Process for HER returns processed observations, the target and the augmented target

processGace' :: Tensor -> Info -> Tensor

Standardize state over all parallel envs

processGace :: Tensor -> Info -> Tensor

Process / Sanitize the Observations from GACE

scaleRewards :: Tensor -> Float -> Tensor

Scale reward to center

sanatizeJSON :: Char -> Char

Sanatize JSON for MLFlow: Names may only contain alphanumerics, underscores (_), dashes (-), periods (.), spaces ( ), and slashes (/).

data Tracker

Data Logging to MLFlow Trackign Server

Constructors

Tracker

Fields

uri :: TrackingURI

Tracking Server URI

experimentId :: ExperimentID

Experiment ID

experimentName :: String

Experiment Name

runIds :: Map String RunID

Run IDs

runId :: Tracker -> String -> RunID

Retrieve a run ID

mkTracker :: TrackingURI -> String -> IO Tracker

Make new Tracker given a Tracking Server URI

mkTracker' :: String -> Int -> String -> IO Tracker

Make new Tracker given a Hostname and Port

newExperiment :: Tracker -> String -> IO Tracker

Create a new Experiment with rng suffix

newExperiment' :: Tracker -> String -> IO Tracker

Create a new Experiment

newRuns :: Tracker -> [String] -> [Param] -> IO Tracker

Create a new run with a set of given paramters

newRuns' :: Int -> Tracker -> IO Tracker

New run with algorithm id and #envs as log params

endRun :: String -> Tracker -> IO Tracker

End a run

endRuns :: Tracker -> IO Tracker

End all runs of a Tracker

endRuns' :: Tracker -> IO ()

End all runs and discard tracker

trackLoss :: Tracker -> Int -> String -> Float -> IO (Response ByteString)

Write Loss to Tracking Server

trackReward :: Tracker -> Int -> Tensor -> IO ()

Write Reward to Tracking Server

filterPerformance :: Map Int (Map String Float) -> [String] -> Map Int (Map String Float)

Filter Performance of all envs

trackEnvState :: Tracker -> HymURL -> Int -> IO ()

Write Current state of the Environment to Trackign Server